add oneOf optimization - 20s - now faster than master
