+++ scratch.js
@@ -5,5 +5,10 @@ const nlp = require('./src/index')
 
 // nlp(`okay, do not use reverse psychology`).debug()
 // nlp(`April, June, and September`).debug()
-nlp(`really mark. tony is cool. u r okay? April, June, and September`).debug()
 // nlp(`among them the ones at HUD`).debug()
+
+// nlp(`really mark. tony is cool. u r okay? April, June, and September`).debug()
+nlp
+  .tokenize('two')
+  .match('(one|two|three)')
+  .debug()
