+++ scripts/test/speed/_performance.json
@@ -7,29 +7,29 @@
     788,
     24453,
     27712,
-    36354,
+    36355,
     39577,
     39660,
     54897
   ],
   "y": [
-    4.8,
+    4.2,
     3.8,
-    6.4,
-    17.4,
-    32.2,
-    527,
-    546.6,
-    575.6,
-    743.6,
-    666,
-    1303.2
+    5.8,
+    12.6,
+    21.4,
+    478.2,
+    466.6,
+    445.2,
+    618.4,
+    527.4,
+    1081.8
   ],
   "key": {
     "x": "Length",
     "y": "Time"
   },
-  "diff": "5.81818",
-  "regression": "0.02015X - -7.94028",
-  "average": "402.41818"
+  "diff": "1.56363",
+  "regression": "0.01658X - -4.49152",
+  "average": "333.21818"
 }
